backend 1.22.3

+++ src/main/java/com/example/shopjava/repos/impl/UtilsImpl.java
@@ -1,8 +1,8 @@
 package com.example.shopjava.repos.impl;
 
-import com.example.shopjava.entities.Laptop;
-import com.example.shopjava.entities.Phone;
-import com.example.shopjava.entities.Watch;
+import com.example.shopjava.entities.product.Laptop;
+import com.example.shopjava.entities.product.Phone;
+import com.example.shopjava.entities.product.Watch;
 import com.example.shopjava.repos.Utils;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
@@ -15,28 +15,28 @@ public class UtilsImpl implements Utils {
     @Override
     public Phone min(List<Phone> phoneList) {
         if (!phoneList.isEmpty())
-            return phoneList.stream().min(Comparator.comparingDouble(Phone::getPrice)).get();
+            return phoneList.stream().min(Comparator.comparingInt(Phone::getPrice)).get();
         return null;
     }
 
     @Override
     public Phone max(List<Phone> phones) {
         if (!phones.isEmpty())
-            return phones.stream().max(Comparator.comparingDouble(Phone::getPrice)).get();
+            return phones.stream().max(Comparator.comparingInt(Phone::getPrice)).get();
         return null;
     }
 
     @Override
     public Laptop maxLaptop(List<Laptop> laptops) {
         if (!laptops.isEmpty())
-            return laptops.stream().max(Comparator.comparingDouble(Laptop::getPrice)).get();
+            return laptops.stream().max(Comparator.comparingInt(Laptop::getPrice)).get();
         return null;
     }
 
     @Override
     public Watch maxWatch(List<Watch> watches) {
         if (!watches.isEmpty())
-            return watches.stream().max(Comparator.comparingDouble(Watch::getPrice)).get();
+            return watches.stream().max(Comparator.comparingInt(Watch::getPrice)).get();
         return null;
     }
